+++ src/data.js
@@ -1,0 +1,7 @@
+const arg = {
+    userInput: {
+        lastName: 'lastName1',
+        firstName: 'firstName1'
+    }
+}
+module.exports = { arg }
